feat: add worker status endpoint
