+++ proto/worker_status.proto
@@ -1,0 +1,30 @@
+// Specification of the user facing gRPC API.
+syntax = "proto3";
+package status;
+
+service StatusApi {
+    // Returns the status of the worker.
+    rpc Status(StatusRequest) returns (StatusResponse) {}
+}
+
+// Request message for the status of the worker.
+message StatusRequest {}
+
+// Response message containing the status of the worker.
+message StatusResponse {
+    // Whether the worker is ready to generate proofs.
+    bool ready = 1;
+    // The version of the worker.
+    string version = 2;
+    // The supported proof types.
+    SupportedProofTypes supported_proof_types = 3;
+}
+
+message SupportedProofTypes {
+    // Whether the worker supports transaction proofs.
+    bool transaction = 1;
+    // Whether the worker supports batch proofs.
+    bool batch = 2;
+    // Whether the worker supports block proofs.
+    bool block = 3;
+}
